requirements.txt
babel==2.17.0
bitarray==3.6.1
bitstring==4.3.1
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
docopt==0.6.2
esptool==5.0.2
fastapi==0.116.2
h11==0.16.0
idna==3.10
iniconfig==2.1.0
intelhex==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==25.0
pipreqs==0.4.13
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pyserial==3.5
pytest==8.4.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
reedsolo==1.7.0
requests==2.32.4
rich==14.1.0
rich-click==1.8.9
simple-websocket==1.1.0
typing_extensions==4.14.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
wsproto==1.2.0
yarg==0.1.10
pytest
mypy
ruff
babel
werkzeug
jinja2
python-multipart
